requirements.txt
django>=1.5
cassandra-driver
cqlengine>=0.20.0
